Пятый коммит: Создание кнопки для смены темы и обработка ошибок для delete_task

+++ main.py
@@ -1,9 +1,12 @@
+# Импорт библиотек для работы с customtkinter и отсальными
 import customtkinter as ctk
-import tkinter as tk
+import tkinter as tk 
 from tkinter import messagebox
 import db
+from PIL import Image, ImageTk
 
 class ToDoApp(ctk.CTk):
+    # Иницализация
     def __init__(self):
         super().__init__()
         ctk.set_appearance_mode("dark")
@@ -63,8 +66,23 @@ class ToDoApp(ctk.CTk):
         self.btn_clear_completed_tasks = ctk.CTkButton(self, text='Очистить выполненные', command=self.clear_completed_tasks, width=120, height=40)
         self.btn_clear_completed_tasks.place(x=700, y=400)
         
+        # Кнопка для возврата задачи
         self.btn_return = ctk.CTkButton(self, text='Вернуть в список', command=self.return_to_tasks, width=120, height=40)
         self.btn_return.place(x=700, y=450)
+        
+        # Кнопка для смены темы
+        
+        light_theme = Image.open('images/light_icon.png')
+        dark_theme = Image.open('images/dark_icon.png')
+        
+        light_theme = light_theme.resize((20, 20), Image.Resampling.LANCZOS)
+        dark_theme = dark_theme.resize((20, 20), Image.Resampling.LANCZOS)
+        
+        self.light_photo = ImageTk.PhotoImage(light_theme)
+        self.dark_photo = ImageTk.PhotoImage(dark_theme)
+        
+        self.btn_theme = ctk.CTkButton(self, image=self.dark_photo, text='', width=20, height=35, hover_color='#333333', command=self.toggle_theme)
+        self.btn_theme.place(x=850, y=10)
     
     # Функция для добавления задачи в базу данных также в listbox
     def add_task(self):
@@ -85,13 +103,17 @@ class ToDoApp(ctk.CTk):
     # Функция для удаления задачи из базы данных также из listbox
     def delete_task(self):
         try:
-            selected_task_index = self.listbox.curselection()[0]
-            full_text = self.listbox.get(selected_task_index)
-            title, description = full_text.split(': ', 1)
-            db.delete_task_from_db(title, description)
-            self.listbox.delete(selected_task_index)
-        except:
+            if self.listbox.curselection():
+                if messagebox.askyesno('Подтверждение', 'Удалить задачу?'):
+                    selected_task_index = self.listbox.curselection()[0]
+                    full_text = self.listbox.get(selected_task_index)
+                    title, description = full_text.split(': ', 1)      
+                    db.delete_task_from_db(title, description)
+                    self.listbox.delete(selected_task_index)
+        except IndexError:
             messagebox.showwarning('Ошибка', 'Выберите задачу для удаления')
+        except Exception as e:
+            messagebox.showerror('Ошибка', f'Произошла ошибка: {e}')
     
     # Функция для добавления задачи в выполненных    
     def add_to_completed(self):
@@ -169,6 +191,14 @@ class ToDoApp(ctk.CTk):
                 self.listbox_completed.insert(tk.END, text)
         except Exception as e:
             messagebox.showerror('Ошибка', f'Не удалось загрузить задачи: {e}')
+    
+    def toggle_theme(self):
+        if ctk.get_appearance_mode() == 'Light':
+            ctk.set_appearance_mode("Dark")
+            self.btn_theme.configure(image=self.dark_photo)
+        else:
+            ctk.set_appearance_mode("Light")
+            self.btn_theme.configure(image=self.light_photo)
             
 app = ToDoApp()
 app.mainloop()
